+++ AppboyKit/ABKLocationManagerProvider.m
@@ -1,13 +1,13 @@
 #import "ABKLocationManagerProvider.h"
 
-#if !TARGET_OS_TV && !defined(ABK_DISABLE_LOCATION_SERVICES)
+#if !TARGET_OS_TV
 #import <CoreLocation/CoreLocation.h>
 #endif
 
 @implementation ABKLocationManagerProvider
 
 + (BOOL)locationServicesEnabled {
-#if !TARGET_OS_TV && !defined(ABK_DISABLE_LOCATION_SERVICES)
+#if !TARGET_OS_TV
   return YES;
 #endif
   return NO;