+++ AppboyKit/headers/AppboyKitLibrary/ABKContentCard.h
@@ -89,7 +89,6 @@ NS_ASSUME_NONNULL_BEGIN
 /*!
  * Manually log a click to Braze for the card.
  * This should only be used for custom contentcard view controllers.
- * The SDK will only log a card click when the card has the url property with a valid url.
  */
 - (void)logContentCardClicked;