+++ Example/Stopwatch/AboutViewController.h
@@ -1,0 +1,8 @@
+#import <UIKit/UIKit.h>
+
+@interface AboutViewController : UIViewController
+
+@property (weak, nonatomic) IBOutlet UILabel *apiKeyLabel;
+@property (weak, nonatomic) IBOutlet UILabel *apiEndpointLabel;
+
+@end