+++ HelloSwift/HelloSwiftNotificationContentExtension/NotificationViewController.swift
@@ -1,0 +1,26 @@
+import AppboyPushStory
+import UIKit
+import UserNotifications
+import UserNotificationsUI
+
+class NotificationViewController: UIViewController, UNNotificationContentExtension {
+
+  @IBOutlet weak var storiesView: ABKStoriesView!
+  var dataSource: ABKStoriesViewDataSource?
+    
+  func didReceive(_ notification: UNNotification) {
+    dataSource = ABKStoriesViewDataSource(notification: notification, storiesView: storiesView, appGroup: "group.Appboy.HelloSwift")
+  }
+    
+  func didReceive(_ response: UNNotificationResponse, completionHandler completion: @escaping (UNNotificationContentExtensionResponseOption) -> Void) {
+    if dataSource != nil {
+      let option: UNNotificationContentExtensionResponseOption = dataSource!.didReceive(response)
+      completion(option)
+    }
+  }
+    
+  override func viewWillDisappear(_ animated: Bool) {
+    dataSource?.viewWillDisappear()
+    super.viewWillDisappear(animated)
+  }
+}